Fix issue with incorrect setup of audio examples (#75)
This fixes a breaking change from BeautifulSoup4 that disallows setting the element `contents[]` to raw text representing a new html tag. The correct way to add a new child element to a tag is to use `el.append()`.

I believe this change occured in v4.11.2 or newer. Recent changes to the project include new ways of handling `.string()` and `.get_text()`.