Improve Effect Details Chapter (#63)
The descriptions for "every effect explained in detail" have been edited
for readability and to be more beginner-friendly. Also the audio
examples have been improved to be more obvious, with a wet and dry
version. Improved code examples for each effect are now located at the
top of each section that match the audio examples.

---------

Co-authored-by: heerman <lwheerman@gmail.com>
Co-authored-by: Ian Clester <ijc@ijc8.me>